% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cat.R
\name{filter_cat}
\alias{filter_cat}
\alias{\%get\%}
\alias{show_categories}
\title{Select variables based on categories}
\usage{
d \%get\% var.cat

show_categories(d)
}
\arguments{
\item{d}{dataset from data packages}

\item{var.cat}{variable category}
}
\description{
All variables belonging to the same category are selected from a dataset
using the \verb{\%get\%} infix operator. Categories are stored in the
"var_data_category" attribute, if it exists. Use \code{show_categories()} to see
all available variable categories.
}
\author{
Samuel Leung, Derek Chiu
}
