% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_native_type.R
\name{to_native_type}
\alias{to_native_type}
\alias{to_native_type.data.frame}
\alias{to_native_type.bccaEndometrial}
\alias{to_native_type.deEndometrial}
\alias{to_native_type.caEndometrial}
\alias{to_native_type.CBOCOUv5}
\title{Change variables back to their native types}
\usage{
to_native_type(x)

\method{to_native_type}{data.frame}(x)

\method{to_native_type}{bccaEndometrial}(x)

\method{to_native_type}{deEndometrial}(x)

\method{to_native_type}{caEndometrial}(x)

\method{to_native_type}{CBOCOUv5}(x)
}
\arguments{
\item{x}{dataset from data packages}
}
\value{
dataset with variables converted back to native types
}
\description{
Takes data package objects and converts variables back to native types
(e.g. numeric, character, factor, Date).
}
\author{
Derek Chiu
}
