% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeconv-package.R
\docType{package}
\name{typeconv-package}
\alias{typeconv}
\alias{typeconv-package}
\title{typeconv: Convert data package variables to their native types}
\description{
Convert variables from clinical data packages to their native types such as character, factor, numeric, and Date.
}
\author{
\strong{Maintainer}: Derek Chiu \email{dchiu@bccrc.ca}

Authors:
\itemize{
  \item Samuel Leung \email{Samuel.Leung@vch.ca}
}

}
\keyword{internal}
